refactor(Solver): include AbstracKnitroProblem in AbstractKnitroSolver
